requirements.txt
requests>=2.31,<3
pyyaml>=6.0,<7
yt-dlp>=2024.1
